+++ crypto/keccak256.h
@@ -43,7 +43,7 @@ extern "C" {
 
 
 void keccak_init(SHA3_CTX *ctx);
-void keccak_update(SHA3_CTX *ctx, const unsigned char *msg, uint16_t size);
+void keccak_update(SHA3_CTX *ctx, const unsigned char *msg, uint32_t size);
 void keccak_final(SHA3_CTX *ctx, unsigned char* result);
